fixed toolbar functionality
