+++ app/src/main/java/edu/quinnipiac/ser210/assignment2part2piano/ChordsActivity.java
@@ -16,7 +16,7 @@ import java.util.Collections;
 import java.util.Locale;
 import java.util.Random;
 
-public class ChordsActivity extends AppCompatActivity implements ChordListFragment.Listener{
+public class ChordsActivity extends AppCompatActivity{
     private RecyclerView mRecyclerView;
     private LinearLayoutManager linearLayoutManager;
     private ArrayList<Chord> mChordData;
@@ -30,77 +30,6 @@ public class ChordsActivity extends AppCompatActivity implements ChordListFragme
         ChordListFragment frag = (ChordListFragment) getSupportFragmentManager().findFragmentById(R.id.fragment_chord_list);
         mChordData = (ArrayList<Chord>) getIntent().getExtras().get("chords");
         frag.setChordData(mChordData);
-    }
-
-    //Gets the menu options from the menu xml and displays them on the tool bar
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu){
-        getMenuInflater().inflate(R.menu.main_menu, menu);
-
-        MenuItem searchItem = menu.findItem(R.id.action_search);
-        SearchView searchView = (SearchView) searchItem.getActionView();
-
-        //Listener for the search view
-        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-            @Override
-            public boolean onQueryTextSubmit(String s) {
-                return false;
-            }
-
-            @Override
-            public boolean onQueryTextChange(String s) {
-                filter(s);
-                return false;
-            }
-        });
-
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean toolbarItemClicked(MenuItem item) {
-        int id = item.getItemId();
-        switch(id){
-            case R.id.action_shuffle:{
-                shuffle();
-                return true;
-            }
-            case R.id.action_random:{
-                random();
-                return true;
-            }
-            default:
-                return false;
-        }
-    }
-
-    //Searches for a chord using a given string from the Search View
-    public void filter(String text){
-        ArrayList<Chord> filteredList = new ArrayList<>();
-
-        for(Chord item : mChordData){
-            if(item.getRoot().toLowerCase().contains(text.toLowerCase(Locale.ROOT))){
-                filteredList.add(item);
-            }
-        }
-
-        if(filteredList.isEmpty()){
-            Toast.makeText(this, "No Chords Found", Toast.LENGTH_SHORT).show();
-        }else{
-            mChordAdapter.filterList(filteredList);
-        }
-    }
-
-    //Shuffles the recycle view
-    public void shuffle(){
-        Collections.shuffle(mChordData, new Random());
-        mChordAdapter.filterList(mChordData);
-    }
-
-    //Scrolls to a random position in the recycle view, resulting a random chord being shown
-    public void random(){
-        int pos = (int) (Math.random() * mChordData.size());
-        linearLayoutManager.scrollToPositionWithOffset(pos, 0);
     }
 
 }
